+++ changelog.md
@@ -21,6 +21,7 @@
 - new widget event "exit" allows stopping of functions before page change (e.g. deleting timers)
 - new widget basic.roundslider with flexible design options
 - new widget device.rtrslider for slider control of set temperature and functionality of device.rtr 
+- new widget status.toast to display notifications 
 
 ### Other New Features
 - bash script "setpermissions" added for setting file permissions during initial setup. To be called by 'sudo bash setpermissions'